Refactoring - part 35
